Re-structured logic of AlarmManager and it's notifications
Since each alarmManager need a unique requestCode, so i added 5 column in the tanamankuItemModels which contains UUID.random.hashCode.absolute so each tanamankuItem has 5 unique IDs that can be used as a requestCode for alarmManager.

Also, in the AlarmManager implementation, i do re-structuring of the logic which will made future development easier and prevent each AlarmManager overwriting each other. And also cancelAlarm() has been tested succesfully.

The last one, for notifications, i have implemented where the notifications not going to be dismissed until user tap on it, by set the fullscreenIntent to empty Intent() (thanks to this reference : https://stackoverflow.com/questions/41639846/android-heads-up-notification-disappears-after-a-few-seconds)